WIP: Use a Config scheme that doesn't depend on nulls
This requires using a graph data structure to store the config values (and record which ones are set), rather than just an instance of the config object.

This also requires parsing input as expressions that define a lookup into the graph rather than functions that can operate on the config object directly.