Dual build for monomac and xamarin-mac
We probably want to publish as xamarin, but currently it's impossible to run unit tests on it. Monomac has almost the same API so with some namespace finagling we can test with monomoc and publish with xamarin. Obviously that's not ideally in terms of testing against what we're actually running, but it's good enough for now.