Improve size calculation
When custom drawing the image, the size calculation did just remove some
constant amount from the image space height, when drawing with a label.
It also did not account for square images (in which case nothing was
removed).

It now calculates the scale for each dimension and uses the smaller one.
When calculating the scale it deducts the text height from the available
height, if any text is visible.