Use PDF storage when importing PDFs with OCR text
This allows OCR text to be maintained when exporting without needing to re-run OCR or risking changing fonts etc if the PDF wasn't created by NAPS2.

Rendering may be slightly less efficient but shouldn't be too bad as we can use the fast path for extracting an image (if we couldn't then we'd already be using PDF storage anyway).